Add javadoc comments to most classes and methods. Group classes into respective packages
