--- src/main/java/Event.java
@@ -1,15 +1,0 @@
-public class Event extends Task{
-    private final String startDate;
-    private final String endDate;
-
-    public Event(String taskName, String startDate, String endDate) {
-        super(taskName);
-        this.startDate = startDate;
-        this.endDate = endDate;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[E]%s (from: %s to: %s)", super.toString(), startDate, endDate);
-    }
-}
