--- src/main/java/Parser.java
@@ -1,5 +1,0 @@
-public class Parser {
-    public static CommandType getCommandType(String input) {
-        return CommandType.getCommandType(input);
-    }
-}
