--- src/main/java/ToDo.java
@@ -1,10 +1,0 @@
-public class ToDo extends Task{
-    public ToDo(String name) {
-        super(name);
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[T]%s", super.toString());
-    }
-}
