+++ src/main/java/command/CommandType.java
@@ -1,3 +1,8 @@
+package command;
+
+/**
+ * Enum representing the different types of commands.
+ */
 public enum CommandType {
     BYE("bye"),
     DEADLINE("deadline"),
@@ -11,10 +16,19 @@ public enum CommandType {
 
     public final String input;
 
+    /**
+     * Constructor for CommandType enum.
+     * @param input the input string
+     */
     CommandType(String input) {
         this.input = input;
     }
 
+    /**
+     * Returns the command type of the input.
+     * @param input the input string
+     * @return the command type of the input
+     */
     public static CommandType getCommandType(String input) {
         for (CommandType type : CommandType.values()) {
             if (type.input.equals(input)) {
