+++ src/main/java/task/Deadline.java
@@ -1,6 +1,16 @@
-public class Deadline extends Task{
+package task;
+
+/**
+ * Deadline class for tasks with deadlines.
+ */
+public class Deadline extends Task {
     private final String deadline;
 
+    /**
+     * Constructor for Deadline class.
+     * @param taskName the name of the task
+     * @param deadline the deadline for the task
+     */
     public Deadline(String taskName, String deadline) {
         super(taskName);
         this.deadline = deadline;
